requirements.txt
streamlit
crewai
python-docx
python-dotenv
langchain
langchain-community
groq